Cleanup, added progress indicator
