+++ compose/src/main/java/de/jensklingenberg/jetpackcomposeplayground/samples/ButtonDemo.kt
@@ -1,6 +1,23 @@
-package de.jensklingenberg.jetpackcomposeplayground.sample.buttondemo
+/*
+ * Copyright 2019 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.jensklingenberg.jetpackcomposeplayground.samples
 
 import android.util.Log
+import androidx.ui.core.CraneWrapper
 import androidx.ui.core.Text
 import androidx.ui.core.dp
 import androidx.ui.core.withDensity
@@ -16,18 +33,15 @@ import androidx.ui.material.borders.BorderSide
 import androidx.ui.material.borders.RoundedRectangleBorder
 import androidx.ui.material.themeColor
 import androidx.ui.material.themeTextStyle
-
+import androidx.ui.graphics.Color
 import androidx.compose.Composable
 import androidx.compose.unaryPlus
-import androidx.ui.core.CraneWrapper
-import androidx.ui.graphics.Color
-
+import androidx.compose.composer
 
 @Composable
 fun ButtonDemo() {
     val onClick: () -> Unit = { Log.e("ButtonDemo", "onClick") }
     CraneWrapper {
-
         MaterialTheme {
             Center {
                 Column(mainAxisAlignment = MainAxisAlignment.SpaceEvenly) {
@@ -37,8 +51,8 @@ fun ButtonDemo() {
                     Button(
                         onClick = onClick,
                         color = +themeColor { secondary },
-                        text = "SECONDARY COLOR"
-                    )
+                        text = "SECONDARY COLOR")
+
                     val outlinedShape = +withDensity {
                         RoundedRectangleBorder(
                             side = BorderSide(Color(0xFF888888.toInt())),
@@ -50,7 +64,9 @@ fun ButtonDemo() {
                             )
                         )
                     }
+
                     TransparentButton(onClick = onClick, shape = outlinedShape, text = "OUTLINED")
+
                     val customColor = Color(0xFFFFFF00.toInt())
                     Button(
                         onClick = onClick,
@@ -61,6 +77,7 @@ fun ButtonDemo() {
                             Text(text = "CUSTOM BUTTON!")
                         }
                     }
+
                     // TODO(Andrey): Disabled button has wrong bg and text color for now.
                     // Need to figure out where will we store their styling. Not a part of
                     // MaterialColors right now and specs are not clear about this.
