+++ compose/src/main/java/de/jensklingenberg/jetpackcomposeplayground/samples/CounterModelDemo.kt
@@ -1,6 +1,7 @@
-package de.jensklingenberg.jetpackcomposeplayground.sample
+package de.jensklingenberg.jetpackcomposeplayground.samples
 
 import androidx.compose.Composable
+import androidx.compose.state
 import androidx.compose.unaryPlus
 import androidx.ui.core.CraneWrapper
 import androidx.ui.core.Text
@@ -9,11 +10,15 @@ import androidx.ui.layout.Column
 import androidx.ui.material.Button
 import androidx.ui.material.MaterialTheme
 import androidx.ui.material.themeTextStyle
-import de.jensklingenberg.jetpackcomposeplayground.ui.CounterModel
 
+
+class CounterModel {
+    val counter = +state { 0 }
+    var header = "Counter demo"
+}
 
 @Composable
-fun Counter() {
+fun CounterModelDemo() {
     CraneWrapper {
 
         MaterialTheme {
