+++ compose/src/main/java/de/jensklingenberg/jetpackcomposeplayground/samples/TextDemo.kt
@@ -1,4 +1,4 @@
-package de.jensklingenberg.jetpackcomposeplayground.sample.text
+package de.jensklingenberg.jetpackcomposeplayground.samples
 
 import androidx.compose.Composable
 import androidx.compose.unaryPlus
@@ -7,10 +7,8 @@ import androidx.ui.core.Text
 import androidx.ui.material.MaterialTheme
 import androidx.ui.material.themeTextStyle
 
-
 @Composable
-
-fun SimpleComposable() {
+fun TextDemo() {
     CraneWrapper {
         MaterialTheme {
             Text(text = "Hello World!", style = +themeTextStyle { h1 })
